+++ packages/braintree-utils/src/braintree.ts
@@ -539,7 +539,7 @@ export interface BraintreeConnectIdentity {
 }
 
 export interface BraintreeConnectLookupCustomerByEmailResult {
-    customerId?: string;
+    customerContextId?: string;
 }
 
 export interface BraintreeConnectAuthenticationOptions {
@@ -579,7 +579,6 @@ export enum BraintreeConnectAuthenticationState {
     SUCCEEDED = 'succeeded',
     FAILED = 'failed',
     CANCELED = 'canceled',
-    UNRECOGNIZED = 'unrecognized', // Info: this is a custom authentication state used on BC side only for users without PayPal account
 }
 
 export interface BraintreeConnectAuthenticationCustomerResult {